backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=24.3
python-dotenv>=1.0
drf-nested-routers>=0.93
mysqlclient>=2.2 ; platform_system != 'Windows'
PyMySQL>=1.1 ; platform_system == 'Windows'
